Add new release for vcu driver which has several optimizations
It has below optimizations and changes at encoder and decoder.

Encoder :
Many intermediate buffers which were allocated and then handed
to the driver are now allocated as a big chunk of memory at
channel creation for the firmware to use.
This reduce memory usage and memory waste because of page alignments.

It also simplifies the encode one frame ioctl as only the source
buffer and the qp table buffer if necessary need to be given
instead of a big number of buffers.

The channel creation ioctl is now given opaque parameters as
the driver doesn't need to know these parameters
(the firmware does) The structure that should be passed in this
opaque parameter is the AL_TEncChanParam.

As the dpb process is now done in the firmware,
two new debug ioctl have been added to retrieve the reconstructed picture.
(AL_MCU_GET_REC_PICTURE and AL_MCU_RELEASE_REC_PICTURE)

AL_MCU_PUT_STREAM_BUFFER which is used with AL_MCU_ENCODE_ONE_FRM.
One gives the buffer which will be filled with the encoded bitstream
of the ENCODE_ONE_FRM frame given.
This buffer will be retrieved using AL_MCU_WAIT_FOR_STATUS.

Decoder :
Most of the ioctl parameters became opaque parameters
if they weren't used by the driver and only by the firmware.

The structure that should be passed in this opaque parameters
are AL_TDecChanParam for al5_channel_config, AL_TDecPictParam and
AL_TDecPicBufferAddrs for al5_decode_msg and AL_TScParam and
AL_TScBufferAddrs for al5_search_sc_msg.

Signed-off-by: Kevin Grandemange <kevin.grandemange@allegrodvt.com>
Signed-off-by: Devarsh Thakkar <devarsht@xilinx.com>