Do not handle interrupts which weren't generated by the ip
Check the AL5_MCU_IRQ_STA register to check if the
interrupt is coming from the device.
Make sure that the interrupt is acknowledged at the
end of the hardirq handler to not get "ghost" interrupts.
(coming from the device but with AL5_MCU_IRQ_STA == 0 because
the level interrupt wasn't really acknowledged at the
end of the hardirq handler)

Signed-off-by: Kevin Grandemange <kevin.grandemange@allegrodvt.com>
Signed-off-by: Devarsh Thakkar <devarsht@xilinx.com>