common: Round page table end address to nearest multiple of 0x40000000
Memory Hotplug API requires a block alignment of 0x40000000. It may not
always be feasible to allocate chunks of 0x40000000.

This patch works around this restriction by allowing the reserved memory
pool to be any size. However, when we send the page table to the Memory
Hotplug API, we round the provided end address to be aligned to
0x40000000. This won't create any issues because the actual allocated
reserved memoy pool is smaller so the vcu will never access this padded
memory.

Signed-off-by: Dylan Yip <dylan.yip@xilinx.com>
Signed-off-by: Jeegar Patel <jeegar.patel@xilinx.com>

Acked-for-series:Varunkumar Allagadapa <varunkumar.allagadapa@xilinx.com>