fix: channel destruction
if channel is created while a user is already destroyed (i.e. Ctrl+C),
the message may arrive later.
We need to check if the creation channel message is in the mailbox before destroying the user.
If it is, destroy the channel. Otherwise the case is already handled

Signed-off-by: Parth Gajjar <parth.gajjar@amd.com>

Acked-for-series: Varunkumar Allagadapa <varunkumar.allagadapa@amd.com>