dmaproxy: Use DMA_TO_DEV in dma_map_page API for destination buffer
linux upsteam commit:
370645f41e6e("dma-mapping: force bouncing if the kmalloc() size...")
For direct DMA, if the size is small enough to have originated from a
kmalloc() cache below ARCH_DMA_MINALIGN, check its alignment against
dma_get_cache_alignment() and bounce if necessary.

In this case encoder is writing some small chunks that is less
than ARCH_DMA_MINALIGN and it's not aligned with
dma_get_cache_alignment()(128) due to this dma_map_page API is bouncing and
gives wrong destiantion address which cause Encoded buffer is corrupted.

This small chunks are part of main buffer that is allocated using
dma_alloc_coherent which is not the kmalloc() buffers and it's
dma aligned buffers but due to above upstream commit it's treating
all the buffer as kmalloc() that is passed in dma_map_page as there
is no check to confirm if the upcoming buffer is allocated using
kmalloc() or not. Planning to raise this query to public forum to
handle this kind of cases.

According to upstream commit kmalloc() buffers are DMA-safe
irrespective of size if the device is coherent or the direction is
DMA_TO_DEVICE (non-desctructive cache maintenance and benign cache
line evictions)

As a workaround Changing the direction in dma_map_page to
DMA_TO_DEV in case of Destination buffer to avoid the force bounce
due to small size DMA transfer and was resulting into wrong address.

Signed-off-by: Yash Ladani <yash.ladani@amd.com>
Signed-off-by: Rohit Visavalia <rohit.visavalia@amd.com>
Acked-by: Vishal Sagar <vishal.sagar@amd.com>
Acked-by: Varunkumar Allagadapa <varunkumar.allagadapa@amd.com>